requirements.txt
joblib~=1.4.2
streamlit~=1.35.0
python-dotenv~=1.0.1
pandas~=2.2.2
matplotlib~=3.9.0
seaborn~=0.13.2
fpdf~=1.7.2
numpy~=1.26.4
pyserial~=3.5
streamviz~=5.1
pillow~=10.3.0
supabase~=2.5.1